+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Allow to use HTML for success content.
+- Use a Stimulus controller value for success content.
 
 ## [3.3.0] - 2022-12-23